After deep investigation, the poor performance on TF2 is caused by some implementation details.

This PR has fixed this issue.

There is the output:

```
[INFO] set data_format = 'channels_last'
initial accuracy: 97.74%
WARNING:tensorflow:AutoGraph could not transform <bound method ModelAdapter._get_grad_xent of <tensorflow.python.eager.function.TfMethodTarget object at 0x7f4c5809fdd8>> and will run it as-is.
Cause: mangled names are not yet supported
To silence this warning, decorate the function with @tf.autograph.experimental.do_not_convert
apgd-ce - 1/10 - 96 out of 1000 successfully perturbed
apgd-ce - 2/10 - 130 out of 1000 successfully perturbed
apgd-ce - 3/10 - 106 out of 1000 successfully perturbed
apgd-ce - 4/10 - 104 out of 1000 successfully perturbed
apgd-ce - 5/10 - 95 out of 1000 successfully perturbed
apgd-ce - 6/10 - 73 out of 1000 successfully perturbed
apgd-ce - 7/10 - 51 out of 1000 successfully perturbed
apgd-ce - 8/10 - 57 out of 1000 successfully perturbed
apgd-ce - 9/10 - 29 out of 1000 successfully perturbed
apgd-ce - 10/10 - 73 out of 774 successfully perturbed
robust accuracy after APGD-CE: 89.60% (total time 55.6 s)
WARNING:tensorflow:AutoGraph could not transform <bound method ModelAdapter.__get_grad_dlr of <tensorflow.python.eager.function.TfMethodTarget object at 0x7f4c58067a90>> and will run it as-is.
Cause: mangled names are not yet supported
To silence this warning, decorate the function with @tf.autograph.experimental.do_not_convert
apgd-dlr - 1/9 - 8 out of 1000 successfully perturbed
apgd-dlr - 2/9 - 20 out of 1000 successfully perturbed
apgd-dlr - 3/9 - 7 out of 1000 successfully perturbed
apgd-dlr - 4/9 - 15 out of 1000 successfully perturbed
apgd-dlr - 5/9 - 4 out of 1000 successfully perturbed
apgd-dlr - 6/9 - 6 out of 1000 successfully perturbed
apgd-dlr - 7/9 - 5 out of 1000 successfully perturbed
apgd-dlr - 8/9 - 7 out of 1000 successfully perturbed
apgd-dlr - 9/9 - 7 out of 960 successfully perturbed
robust accuracy after APGD-DLR: 88.81% (total time 123.2 s)
WARNING:tensorflow:AutoGraph could not transform <bound method ModelAdapter.__get_jacobian of <tensorflow.python.eager.function.TfMethodTarget object at 0x7f4c001f1cf8>> and will run it as-is.
Cause: mangled names are not yet supported
To silence this warning, decorate the function with @tf.autograph.experimental.do_not_convert
fab - 1/9 - 1 out of 1000 successfully perturbed
fab - 2/9 - 4 out of 1000 successfully perturbed
fab - 3/9 - 1 out of 1000 successfully perturbed
fab - 4/9 - 4 out of 1000 successfully perturbed
fab - 5/9 - 3 out of 1000 successfully perturbed
fab - 6/9 - 2 out of 1000 successfully perturbed
fab - 7/9 - 0 out of 1000 successfully perturbed
fab - 8/9 - 0 out of 1000 successfully perturbed
fab - 9/9 - 1 out of 881 successfully perturbed
robust accuracy after FAB: 88.65% (total time 480.6 s)
square - 1/9 - 35 out of 1000 successfully perturbed
square - 2/9 - 34 out of 1000 successfully perturbed
square - 3/9 - 29 out of 1000 successfully perturbed
square - 4/9 - 29 out of 1000 successfully perturbed
square - 5/9 - 22 out of 1000 successfully perturbed
square - 6/9 - 13 out of 1000 successfully perturbed
square - 7/9 - 19 out of 1000 successfully perturbed
square - 8/9 - 12 out of 1000 successfully perturbed
square - 9/9 - 21 out of 865 successfully perturbed
robust accuracy after SQUARE: 86.51% (total time 749.6 s)
max Linf perturbation: 0.30000, nan in tensor: 0, max: 1.00000, min: 0.00000
robust accuracy: 86.51%

```

As you see, that execution time has significant improvement.