@fra31 2020-07-15
It seems to work fine, nice! Could you please just add `@tf.autograph.experimental.do_not_convert` to silence the warnings?