@CNOCycle 2020-07-15
Since there is no clear document which describes this issue on TF official website. 

I'm not able to fix these annoy warning currently.