@fra31 2020-07-15
I tried locally and adding `@tf.autograph.experimental.do_not_convert` right after `@tf.function` does the job. Doesn't it work for you?