@CNOCycle 2020-07-15
Thanks your suggestion. 

I tested it again. My debugger still returned some strange errors after adding the decoration. 

But, it worked properly without my debugger. 

I will give a quick fix.