@fra31 2020-07-15
Ok, merged, thanks for the improvement! I can also add the fix for the warning myself if necessary.